+++ lib/rails_ai_build/rbac.rb
@@ -6,8 +6,8 @@ module RailsAiBuild
     READ_ONLY_BOOST = Tools::Registry::BOOST_TOOL_NAMES
 
     DEFAULT_ROLES = {
-      admin: %i[read_file write_file grep list_files shell] + READ_ONLY_BOOST,
-      developer: %i[read_file write_file grep list_files shell] + READ_ONLY_BOOST,
+      admin: %i[read_file write_file grep list_files shell run_generator host_safety_check] + READ_ONLY_BOOST,
+      developer: %i[read_file write_file grep list_files shell run_generator host_safety_check] + READ_ONLY_BOOST,
       reviewer: %i[read_file grep list_files] + READ_ONLY_BOOST,
       viewer: %i[read_file list_files] + READ_ONLY_BOOST
     }.freeze